+++ .gitignore
@@ -4,4 +4,4 @@
 /backend/coverage/
 /node_modules/
 /backend/qodana.yaml
-/backend/src/conf/.env
+/backend/conf/.env
