+++ backend/api/models/equipamentoSNModel.js
