+++ backend/api/repositories/equipamentosRepository.js
