+++ backend/api/services/equipamentosService.js
