+++ backend/db/equipamentos.json
