+++ backend/di/equipamentoInjector.js
