+++ backend/tests/step_definitions/services/equipamentos/adicionarEquipamentos.services.steps.js
@@ -1,11 +1,11 @@
 const {loadFeature, defineFeature} = require('jest-cucumber');
 const supertest = require ('supertest');
-const app = require('../../../../src/app.js');
-const equipamentoService = require('../../../../src/api/services/equipamentosService');
-const equipamentoInjector = require('../../../../src/di/equipamentoInjector');
-const modelSN = require('../../../../src/api/models/equipamentoSNModel');
-const modelPatrimonio = require('../../../../src/api/models/equipamentoPatrimonioModel');
-const EquipamentosRepository = require('../../../../src/api/repositories/equipamentosRepository.js');
+const app = require('../../../../app.js');
+const equipamentoService = require('../../../../api/services/equipamentosService.js');
+const equipamentoInjector = require('../../../../di/equipamentoInjector');
+const modelSN = require('../../../../api/models/equipamentoSNModel.js');
+const modelPatrimonio = require('../../../../api/models/equipamentoPatrimonioModel.js');
+const EquipamentosRepository = require('../../../../api/repositories/equipamentosRepository.js');
 
 const feature = loadFeature('tests/features/equipamentos/adicionarEquipamento.feature');
 defineFeature(feature, (test) => {
